Fix another conversion warning

+++ source/binding/Expressions_eval.cpp
@@ -367,7 +367,7 @@ optional<ConstantRange> RangeSelectExpression::getRange(const ConstantValue& cl,
     int32_t width = 1;
     if (t.kind == SymbolKind::PackedArrayType) {
         // TODO: handle overflow?
-        width = t.as<PackedArrayType>().elementType.getBitWidth();
+        width = (int32_t)t.as<PackedArrayType>().elementType.getBitWidth();
         msb *= width;
     }
